Move get_cache_block_size to CacheEngine
