+++ cacheflow/model_executor/__init__.py
@@ -1,12 +1,10 @@
 from cacheflow.model_executor.input_metadata import InputMetadata
 from cacheflow.model_executor.model_loader import get_model
-from cacheflow.model_executor.utils import (set_random_seed,
-                                            get_cache_block_size)
+from cacheflow.model_executor.utils import set_random_seed
 
 
 __all__ = [
     "InputMetadata",
-    "get_cache_block_size",
     "get_model",
     "set_random_seed",
 ]
